initial commit
this thing is working. Will be open to feedback on the forums and blog
about what changes would be most useful from here. Most of all though,
there's less than 300 lines of code in the view and the controller.
Worth reading through to see this is not a complicated system!